+++ src/three/3d-space/cameras/front-camera.ts
