+++ src/three/3d-space/cameras/side-camera.ts
