+++ src/three/3d-space/cameras/top-camera.ts
